pyproject.toml
[project]
name = "python-project-83"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "flake8>=7.1.2",
]
